Close should not stop logging. System implementations of TraceListener do close any open resources (e.g. files) when they receive the Close() signal, however they do not stop logging; the next Write() operation will reopen files, etc, as needed.
Serilog does not have the equivalent concept, but I have changed the implementation to at least generate a new reference to Log.Logger, which will release any old reference, and have a similar effect of "refreshing" the trace listener.

Another alternative would be to simply ignore the operation, like it does for Flush(), which also has no equivalent in Serilog.

Either way, the SerilogTraceListener should not stop logging messages when Close() is called.