Make project multi-targeted to net45, net46 and netstandard1.3
Tries to match the available serilog targets as closely as possible.
Serilog also targets netstandard1.0 but System.Diagnostics.TraceSource
which provides TraceListeners is not available for this target.

netstandard1.3 also does not have something comparable to TraceTransfer
so we only conditionally enable it for net45 and net46. Due to this some
other small parts of codes and a test have to become conditional too.

Modify Build.ps1 to use new 'dotnet' tool instead of raw msbuild and nuget.
Also make it execute the tests before creating the nuget package.

Aims to fix issue #19